Refactor and create a posteriors package
