+++ gpflow/conditionals/dispatch.py
@@ -21,9 +21,9 @@ conditional._gpflow_internal_register = conditional.register
 
 conditional.register = deprecated(
     reason="Registering new implementations of conditional() is deprecated. "
-    "Instead, create your own subclass of gpflow.posteriors.AbstractPosterior "
-    "and register an implementation of gpflow.posteriors.get_posterior_class "
-    "that returns your class."
+    "Instead, create your own subclass of gpflow.posteriors.Posterior and "
+    "register an implementation of gpflow.posteriors.get_posterior_class that "
+    "returns your class."
 )(conditional._gpflow_internal_register)
 
 sample_conditional = Dispatcher("sample_conditional")
